# fix transcode's performance problem