This addresses part of https://github.com/bicycle1885/CodecZstd.jl/issues/13.
```
~/w/TranscodingStreams (master|…) $
julia buffering.zstd.jl; and git co transcode; julia buffering.zstd.jl
 24.045589 seconds (42.10 k allocations: 3.086 GiB, 1.03% gc time)
 23.980275 seconds (1.67 k allocations: 3.084 GiB, 1.70% gc time)
 24.218068 seconds (1.67 k allocations: 3.084 GiB, 1.63% gc time)
 24.023841 seconds (1.67 k allocations: 3.084 GiB, 1.61% gc time)
 24.065087 seconds (1.67 k allocations: 3.084 GiB, 1.65% gc time)
Switched to branch 'transcode'
Your branch is up to date with 'origin/transcode'.
 15.628735 seconds (43.12 k allocations: 7.047 GiB, 1.36% gc time)
 16.158169 seconds (30 allocations: 7.045 GiB, 3.93% gc time)
 15.896305 seconds (30 allocations: 7.045 GiB, 3.88% gc time)
 15.976704 seconds (30 allocations: 7.045 GiB, 3.93% gc time)
 15.972673 seconds (30 allocations: 7.045 GiB, 3.86% gc time)
```